feat: Skip wallet onboarding for developers
