+++ test/e2e/default-fixture.js
@@ -1,0 +1,248 @@
+const { NetworkStatus } = require('@metamask/network-controller');
+const { CHAIN_IDS } = require('../../shared/constants/network');
+const { FirstTimeFlowType } = require('../../shared/constants/onboarding');
+
+const FIXTURE_STATE_METADATA_VERSION = 74;
+
+function defaultFixture(inputChainId = CHAIN_IDS.LOCALHOST) {
+  return {
+    data: {
+      AuthenticationController: {
+        isSignedIn: true,
+      },
+      UserStorageController: {
+        isProfileSyncingEnabled: true,
+      },
+      MetamaskNotificationsController: {
+        subscriptionAccountsSeen: [],
+        isFeatureAnnouncementsEnabled: false,
+        isMetamaskNotificationsEnabled: false,
+        isMetamaskNotificationsFeatureSeen: false,
+        metamaskNotificationsList: [],
+        metamaskNotificationsReadList: [],
+      },
+      AccountsController: {
+        internalAccounts: {
+          selectedAccount: 'd5e45e4a-3b04-4a09-a5e1-39762e5c6be4',
+          accounts: {
+            'd5e45e4a-3b04-4a09-a5e1-39762e5c6be4': {
+              id: 'd5e45e4a-3b04-4a09-a5e1-39762e5c6be4',
+              address: '0x5cfe73b6021e818b776b421b1c4db2474086a7e1',
+              metadata: {
+                name: 'Account 1',
+                lastSelected: 1665507600000,
+                keyring: {
+                  type: 'HD Key Tree',
+                },
+              },
+              options: {},
+              methods: [
+                'personal_sign',
+                'eth_sign',
+                'eth_signTransaction',
+                'eth_signTypedData_v1',
+                'eth_signTypedData_v3',
+                'eth_signTypedData_v4',
+              ],
+              type: 'eip155:eoa',
+            },
+          },
+        },
+      },
+      AlertController: {
+        alertEnabledness: {
+          unconnectedAccount: true,
+          web3ShimUsage: true,
+        },
+        unconnectedAccountAlertShownOrigins: {},
+        web3ShimUsageOrigins: {},
+      },
+      AnnouncementController: {
+        announcements: {
+          8: {
+            date: '2021-11-01',
+            id: 8,
+            isShown: false,
+          },
+        },
+      },
+      NetworkOrderController: {
+        orderedNetworkList: [],
+      },
+      AccountOrderController: {
+        pinnedAccountList: [],
+        hiddenAccountList: [],
+      },
+      AppStateController: {
+        browserEnvironment: {},
+        nftsDropdownState: {},
+        connectedStatusPopoverHasBeenShown: true,
+        termsOfUseLastAgreed:
+          '__FIXTURE_SUBSTITUTION__currentDateInMilliseconds',
+        defaultHomeActiveTabName: null,
+        fullScreenGasPollTokens: [],
+        notificationGasPollTokens: [],
+        popupGasPollTokens: [],
+        qrHardware: {},
+        recoveryPhraseReminderHasBeenShown: true,
+        recoveryPhraseReminderLastShown:
+          '__FIXTURE_SUBSTITUTION__currentDateInMilliseconds',
+        showTestnetMessageInDropdown: true,
+        trezorModel: null,
+        usedNetworks: {
+          [CHAIN_IDS.MAINNET]: true,
+          [CHAIN_IDS.LINEA_MAINNET]: true,
+          [CHAIN_IDS.GOERLI]: true,
+          [CHAIN_IDS.LOCALHOST]: true,
+        },
+        snapsInstallPrivacyWarningShown: true,
+      },
+      CurrencyController: {
+        currentCurrency: 'usd',
+        currencyRates: {
+          ETH: {
+            conversionDate: 1665507600.0,
+            conversionRate: 1700.0,
+            usdConversionRate: 1700.0,
+          },
+        },
+      },
+      GasFeeController: {
+        estimatedGasFeeTimeBounds: {},
+        gasEstimateType: 'none',
+        gasFeeEstimates: {},
+      },
+      KeyringController: {
+        vault:
+          '{"data":"WHaP1FrrtV4zUonudIppDifsLHF39g6oPkVksAIdWAHBRzax1uy1asfAJprR7u72t4/HuYz5yPIFQrnNnv+hwQu9GRuty88VKMnvMy+sq8MNtoXI+C54bZpWa8r4iUQfa0Mj/cfJbpFpzOdF1ZYXahTfTcU5WsrHwvJew842CiJR4B2jmCHHXfm/DxLK3WazsVQwXJGx/U71UelGoOOrT8NI28EKrAwgPn+7Xmv0j92gmhau30N7Bo2fr6Zv","iv":"LfD8/tY1EjXzxuemSmDVdA==","keyMetadata":{"algorithm":"PBKDF2","params":{"iterations":600000}},"salt":"nk4xdpmMR+1s5BYe4Vnk++XAQwrISI2bCtbMg7V1wUA="}',
+      },
+      MetaMetricsController: {
+        eventsBeforeMetricsOptIn: [],
+        fragments: {},
+        metaMetricsId: null,
+        participateInMetaMetrics: false,
+        traits: {},
+      },
+      NetworkController: {
+        selectedNetworkClientId: 'networkConfigurationId',
+        networksMetadata: {
+          networkConfigurationId: {
+            EIPS: {},
+            status: NetworkStatus.Available,
+          },
+        },
+        providerConfig: {
+          chainId: inputChainId,
+          nickname: 'Localhost 8545',
+          rpcPrefs: {},
+          rpcUrl: 'http://localhost:8545',
+          ticker: 'ETH',
+          type: 'rpc',
+          id: 'networkConfigurationId',
+        },
+        networkConfigurations: {
+          networkConfigurationId: {
+            chainId: inputChainId,
+            nickname: 'Localhost 8545',
+            rpcPrefs: {},
+            rpcUrl: 'http://localhost:8545',
+            ticker: 'ETH',
+            networkConfigurationId: 'networkConfigurationId',
+          },
+        },
+      },
+      OnboardingController: {
+        completedOnboarding: true,
+        firstTimeFlowType: FirstTimeFlowType.import,
+        onboardingTabs: {},
+        seedPhraseBackedUp: true,
+      },
+      PermissionController: {
+        subjects: {},
+      },
+      PreferencesController: {
+        advancedGasFee: null,
+        currentLocale: 'en',
+        useExternalServices: true,
+        dismissSeedBackUpReminder: true,
+        featureFlags: {},
+        forgottenPassword: false,
+        identities: {
+          '0x5cfe73b6021e818b776b421b1c4db2474086a7e1': {
+            address: '0x5cfe73b6021e818b776b421b1c4db2474086a7e1',
+            lastSelected: 1665507600000,
+            name: 'Account 1',
+          },
+        },
+        ipfsGateway: 'dweb.link',
+        knownMethodData: {},
+        ledgerTransportType: 'webhid',
+        lostIdentities: {},
+        openSeaEnabled: false,
+        preferences: {
+          hideZeroBalanceTokens: false,
+          showExtensionInFullSizeView: false,
+          showFiatInTestnets: false,
+          showTestNetworks: false,
+          smartTransactionsOptInStatus: false,
+          useNativeCurrencyAsPrimaryCurrency: true,
+          petnamesEnabled: true,
+        },
+        selectedAddress: '0x5cfe73b6021e818b776b421b1c4db2474086a7e1',
+        theme: 'light',
+        useBlockie: false,
+        useNftDetection: false,
+        useNonceField: false,
+        usePhishDetect: true,
+        useTokenDetection: false,
+        useCurrencyRateCheck: true,
+        useMultiAccountBalanceChecker: true,
+        useRequestQueue: true,
+      },
+      SelectedNetworkController: {
+        domains: {},
+      },
+      SmartTransactionsController: {
+        smartTransactionsState: {
+          fees: {},
+          liveness: true,
+          smartTransactions: {
+            [CHAIN_IDS.MAINNET]: [],
+          },
+        },
+      },
+      SubjectMetadataController: {
+        subjectMetadata: {
+          'https://metamask.github.io': {
+            extensionId: null,
+            iconUrl: null,
+            name: 'MetaMask < = > Ledger Bridge',
+            origin: 'https://metamask.github.io',
+            subjectType: 'website',
+          },
+        },
+      },
+      TokensController: {
+        allDetectedTokens: {},
+        allIgnoredTokens: {},
+        allTokens: {},
+        detectedTokens: [],
+        ignoredTokens: [],
+        tokens: [],
+      },
+      // TokenRatesController: {
+      //   contractExchangeRates: {},
+      // },
+      TransactionController: {
+        transactions: {},
+      },
+      config: {},
+      firstTimeInfo: {
+        date: 1665507600000,
+        version: '10.21.0',
+      },
+    },
+  };
+}
+
+module.exports = { defaultFixture, FIXTURE_STATE_METADATA_VERSION };
